validated form

+++ frontend/src/App.tsx
@@ -3,6 +3,7 @@ import { useState} from "react";
 
 const App = () => {
   const [activeTab, setActiveTab] = useState("add");
+  const today=new Date().toISOString().split('T')[0];
   const [formData, setformData] = useState({
     name: "",
     employeeId: "",
@@ -39,36 +40,72 @@ async function handlechange(e: React.ChangeEvent<HTMLInputElement | HTMLSelectEl
 }
 
 async function handleSubmit() {
-  const response=await axios.post("https://employee-mangement-f39u.onrender.com/push-user",{
-    name:formData.name,
-    employeeId:formData.employeeId,
-    email:formData.email,
-    phoneNumber:formData.phoneNumber,
-    department:formData.department,
-    dateOfJoining:formData.dateOfJoining,
-    role:formData.role,
-
-  });
-  alert(response.data.message);
-}
-
-async function handleSearch() {
-  const response=await axios.get("https://employee-mangement-f39u.onrender.com/get-user",{
-    params:{search:formData.Searchid},
-  });
-  if(response.data.message!="Employee not found" && response.data.message!="Server error"){
-    setSearchResult(response.data);
+  if(!formData.name){
+    alert("Enter the name feild")
+  }
+  else if(!formData.employeeId){
+    alert("Enter the Employee Id feild")
+  }
+  else if(!formData.email){
+    alert("Enter the Email feild")
+  }
+  else if(!formData.phoneNumber){
+    alert("Enter the Phone Number feild")
+  }
+  else if(!formData.department){
+    alert("Enter the Department feild")
+  }
+  else if(!formData.dateOfJoining){
+    alert("Enter the Date Of Joining feild")
+  }
+  else if(!formData.role){
+    alert("Enter the Role feild")
+  }
+  else if(formData.phoneNumber.length>10){
+    alert("Please Enter 10 Digit Phone Number");
   }
   else{
+    const response=await axios.post("https://employee-mangement-f39u.onrender.com/push-user",{
+      name:formData.name,
+      employeeId:formData.employeeId,
+      email:formData.email,
+      phoneNumber:formData.phoneNumber,
+      department:formData.department,
+      dateOfJoining:formData.dateOfJoining,
+      role:formData.role,
+
+    });
     alert(response.data.message);
   }
 }
 
+async function handleSearch() {
+  if(!formData.Searchid){
+    alert("Enter the Search ID to Search");
+  }
+  else{
+    const response=await axios.get("https://employee-mangement-f39u.onrender.com/get-user",{
+      params:{search:formData.Searchid},
+    });
+    if(response.data.message!="Employee not found" && response.data.message!="Server error"){
+      setSearchResult(response.data);
+    }
+    else{
+      alert(response.data.message);
+    }
+  }
+}
+
 async function handleDelete() {
-  const response = await axios.delete("https://employee-mangement-f39u.onrender.com/del-user",{
-    params:{delid:formData.deleteID},
-  })
-  alert(response.data.message);
+  if(!formData.deleteID){
+    alert("Enter the Delete ID to Delete");
+  }
+  else{
+    const response = await axios.delete("https://employee-mangement-f39u.onrender.com/del-user",{
+      params:{delid:formData.deleteID},
+    })
+    alert(response.data.message);
+  }
 }
 
   return (
@@ -266,6 +303,7 @@ async function handleDelete() {
               }}
               type="date"
               name="dateOfJoining"
+              max={today}
               onChange={handlechange}
               required
             />
@@ -327,7 +365,7 @@ async function handleDelete() {
               <p><strong>Email:</strong> {searchResult.email}</p>
               <p><strong>Phone Number:</strong> {searchResult.phoneNumber}</p>
               <p><strong>Department:</strong> {searchResult.department}</p>
-              <p><strong>Date of Joining:</strong> {searchResult.dateOfJoining}</p>
+              <p><strong>Date of Joining:</strong> {searchResult.dateOfJoining.split("T")[0]}</p>
               <p><strong>Role:</strong> {searchResult.role}</p>
             </div>
           )}
